+++ src/legacy/api/EntityAPI.cpp
@@ -1447,7 +1447,7 @@ Local<Value> EntityClass::quickEvalMolangScript(const Arguments& args) {
     try {
         Actor* actor = get();
         if (!actor) return Local<Value>();
-        return Number::newNumber(actor->quickEvalMolangScript(args[0].toStr()));
+        return Number::newNumber(actor->evalMolang(args[0].toStr()));
     }
     CATCH("Fail in quickEvalMolangScript!");
 }