+++ src/legacy/api/InternationalAPI.cpp
@@ -2,7 +2,7 @@
 #include "api/InternationalAPI.h"
 
 #include <FMT/args.h>
-#include <ll/api/i18n/I18nAPI.h>
+#include <ll/api/i18n/I18n.h>
 
 void FormatHelper(
     std::vector<Local<Value>>&                          args,