+++ src/legacy/legacyapi/command/DynamicCommand.cpp
@@ -2,6 +2,7 @@
 
 #include "fmt/core.h"
 #include "ll/api/Logger.h"
+#include "ll/api/command/CommandRegistrar.h"
 #include "ll/api/memory/Closure.h"
 #include "ll/api/memory/Memory.h"
 #include "ll/api/utils/ErrorUtils.h"
@@ -34,6 +35,7 @@
 #include <algorithm>
 #include <cstddef>
 #include <cstdint>
+#include <ll/api/command/CommandRegistrar.h>
 #include <memory>
 #include <sstream>
 #include <stdexcept>
@@ -42,6 +44,15 @@
 #include <unordered_map>
 #include <utility>
 #include <vector>
+
+template <>
+bool CommandRegistry::parse<std::pair<
+    std::string,
+    int>>(void* p, CommandRegistry::ParseToken const& token, CommandOrigin const&, int, std::string&, std::vector<std::string>&)
+    const {
+    *(std::pair<std::string, int>*)p = std::pair<std::string, int>{token.toString(), (int)getEnumData(token)};
+    return true;
+}
 
 #define logger lse::getSelfPluginInstance().getLogger()
 
@@ -453,30 +464,32 @@ DynamicCommand::preSetup(CommandRegistry& registry, std::unique_ptr<class Dynami
             if (param.type == ParameterType::Enum) {
                 // clone BDS's enum
                 if (commandInstance->enumRanges.count(param.description) == 0) {
-                    auto namesInBds = registry.getEnumNames();
-                    auto iter       = std::find(namesInBds.begin(), namesInBds.end(), param.description);
-                    if (iter == namesInBds.end())
-                        throw std::runtime_error(
-                            "Enum " + std::string(param.description) + "not found in command and BDS"
-                        );
-                    commandInstance->setEnum(*iter, registry.getEnumValues(*iter));
+                    if (!ll::command::CommandRegistrar::getInstance().hasEnum(param.description))
+                        throw std::runtime_error("Enum " + param.description + "not found in command and BDS");
+                    std::vector<std::string> results;
+                    auto&                    enums = registry.mEnums;
+                    auto iter = std::find_if(enums.begin(), enums.end(), [&](CommandRegistry::Enum const& r) {
+                        return r.name == name;
+                    });
+                    if (iter != enums.end())
+                        for (auto& i : iter->values) results.push_back(registry.mEnumValues.at(i.second));
+
+                    commandInstance->setEnum(param.description, results);
                 }
             } else if (param.type == ParameterType::SoftEnum) {
                 // add empty Soft Enum if not found in command and BDS
                 if (commandInstance->softEnums.count(param.description) == 0) {
-                    auto namesInBds = registry.getSoftEnumNames();
-                    auto iter       = std::find(namesInBds.begin(), namesInBds.end(), param.description);
-                    if (iter == namesInBds.end()) commandInstance->setSoftEnum(param.description, {});
+                    if (!ll::command::CommandRegistrar::getInstance().hasSoftEnum(param.description))
+                        commandInstance->setSoftEnum(param.description, {});
                 }
             }
         }
 
         // fix enum name with prefix '_...' if Enum name is exists in BDS
-        auto                                                            namesInBds = registry.getEnumNames();
         std::unordered_map<std::string_view, std::pair<size_t, size_t>> convertedEnumRanges;
         for (auto& [desc, range] : commandInstance->enumRanges) {
             std::string fixedName = desc.data();
-            while (std::find(namesInBds.begin(), namesInBds.end(), fixedName) != namesInBds.end()) {
+            while (ll::command::CommandRegistrar::getInstance().hasEnum(fixedName)) {
                 fixedName.append("_");
             }
             std::string_view fixedView = desc;
@@ -502,12 +515,8 @@ DynamicCommand::preSetup(CommandRegistry& registry, std::unique_ptr<class Dynami
                 values.emplace_back(*iter, index);
                 ++index;
             }
-            registry._addEnumValuesInternal(
-                fixedView.data(),
-                values,
-                ++Bedrock::typeid_t<CommandRegistry>::_getCounter(),
-                &CommandRegistry::parseEnumStringAndInt
-            );
+            registry
+                ._addEnumValuesInternal(fixedView.data(), values, ++Bedrock::typeid_t<CommandRegistry>::_getCounter(), &CommandRegistry::parse<std::pair<std::string, int>>);
         }
         commandInstance->enumRanges.swap(convertedEnumRanges);
 
@@ -526,7 +535,9 @@ DynamicCommand::preSetup(CommandRegistry& registry, std::unique_ptr<class Dynami
         if (!commandInstance->alias_.empty()) registry.registerAlias(commandInstance->name_, commandInstance->alias_);
         auto builder = commandInstance->builder->get();
         for (auto& overload : commandInstance->overloads) {
-            registry.registerOverload(commandInstance->name_, builder, commandInstance->buildOverload(overload));
+            auto& si                                                    = *registry.findCommand(commandInstance->name_);
+            si.overloads.emplace_back(CommandVersion{}, builder).params = commandInstance->buildOverload(overload);
+            registry.registerOverloadInternal(si, si.overloads.back());
         }
         // commandInstance->overloads.clear();
         auto res = dynamicCommandInstances.emplace(commandInstance->name_, std::move(commandInstance));
@@ -628,15 +639,6 @@ std::unique_ptr<class DynamicCommandInstance> DynamicCommand::createCommand(
     }
     command->setCallback(std::move(callback));
     return std::move(command);
-}
-
-bool DynamicCommand::unregisterCommand(CommandRegistry& registry, std::string const& name) {
-    if (registry.unregisterCommand(name)) {
-        dynamicCommandInstances.erase(name);
-        updateAvailableCommands(registry);
-        return true;
-    }
-    return false;
 }
 
 void DynamicCommand::updateAvailableCommands(CommandRegistry& registry) {
@@ -881,8 +883,7 @@ std::string DynamicCommandInstance::setSoftEnum(std::string const& name, std::ve
     if (!hasRegistered()) {
         softEnums.emplace(name, values);
     } else {
-        auto names = registry.getSoftEnumNames();
-        if (std::find(names.begin(), names.end(), name) == names.end()) {
+        if (!ll::command::CommandRegistrar::getInstance().hasSoftEnum(name)) {
             registry.addSoftEnum(name, values);
             return name;
         }
@@ -900,8 +901,7 @@ bool DynamicCommandInstance::addSoftEnumValues(std::string const& name, std::vec
             setSoftEnum(name, values);
         }
     } else {
-        auto names = registry.getSoftEnumNames();
-        if (std::find(names.begin(), names.end(), name) == names.end()) {
+        if (!ll::command::CommandRegistrar::getInstance().hasSoftEnum(name)) {
             registry.addSoftEnum(name, values);
             return true;
         }