+++ CHANGELOG.md
@@ -5,6 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.2] - 2024-02-06
+
+### Fixed
+
+- Fix runcmdEx error when command has no output
+
+## [0.2.1] - 2024-02-06
+
+### Fixed
+
+- Fix endless loop
+
+## [0.2.0] - 2024-02-06
+
+### Added
+
+- Support LeviLamina 0.7.x
+
+### Fixed
+
+- Fix Device::getIP
+
 ## [0.1.6] - 2024-02-05
 
 ### Fixed
@@ -23,8 +45,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First release.
 
-  [0.1.5]: https://github.com/LiteLDev/LegacyScriptEngine/compare/v0.1.4...v0.1.5
-
+[0.2.2]: https://github.com/LiteLDev/LegacyScriptEngine/compare/v0.2.1...v0.2.2
+[0.2.1]: https://github.com/LiteLDev/LegacyScriptEngine/compare/v0.2.0...v0.2.1
+[0.2.0]: https://github.com/LiteLDev/LegacyScriptEngine/compare/v0.1.6...v0.2.0
 [0.1.6]: https://github.com/LiteLDev/LegacyScriptEngine/compare/v0.1.5...v0.1.6
 [0.1.5]: https://github.com/LiteLDev/LegacyScriptEngine/compare/v0.1.4...v0.1.5
 [0.1.4]: https://github.com/LiteLDev/LegacyScriptEngine/releases/tag/v0.1.4